recording button states fixes
