+++ packages/react-native-room-kit/src/components/Header.tsx
@@ -11,6 +11,7 @@ import { HMSManageAudioOutput } from './HMSManageAudioOutput';
 import { HMSRecordingIndicator } from './HMSRecordingIndicator';
 import { CompanyLogo } from './CompanyLogo';
 import { HMSLiveIndicator } from './HMSLiveIndicator';
+import { HMSLiveViewerCount } from './HMSLiveViewerCount';
 
 interface HeaderProps {
   transparent?: boolean;
@@ -41,9 +42,11 @@ export const _Header: React.FC<HeaderProps> = ({
         <View style={styles.logoContainer}>
           <CompanyLogo style={styles.logo} />
 
+          <HMSLiveIndicator />
+
           <HMSRecordingIndicator />
 
-          <HMSLiveIndicator />
+          <HMSLiveViewerCount />
         </View>
 
         {showControls ? (
